Movement/Springs/HurtEffect
Along with 2 new tracks for the music, I updated the player movement when wall jumping to make it a bit more fun to do. After that, I tackled changing the hurt animation from a static Pizzaboss animation into a red flashing that will happen over any other animations (Jumping, running, ect). I also added a new spring prefab that will launch the player into the air when stepped on. It's strength is adjustable in the side bar.