Crouching and Firebars
With this update, I've added the ability to crouch. It's still in its early form, so I'm expecting  to have to fix a lot of bugs and make some tweaks to the way it feels. Apart from that, I also added the new FireBar obstacle with Yomikester's epic fireball textures. It's exactly what you'd think it is.